Use single path to display slider control point connections
